Rework deduplication using lazy imports
