+++ python/hsfs/connection.py
@@ -14,187 +14,17 @@
 #   limitations under the License.
 #
 
-from __future__ import annotations
-
-from typing import Optional
-
-import hopsworks_common.connection
-from hopsworks_common import client, usage, util
 from hopsworks_common.connection import (
     CERT_FOLDER_DEFAULT,
     HOPSWORKS_PORT_DEFAULT,
     HOSTNAME_VERIFICATION_DEFAULT,
+    Connection,
 )
-from hopsworks_common.decorators import connected, not_connected
-from hsfs import engine, feature_store
-from hsfs.core import feature_store_api
 
 
 __all__ = [
-    "Connection",
     "CERT_FOLDER_DEFAULT",
     "HOPSWORKS_PORT_DEFAULT",
     "HOSTNAME_VERIFICATION_DEFAULT",
+    "Connection",
 ]
-
-
-class Connection(hopsworks_common.connection.Connection):
-    """A feature store connection object.
-
-    The connection is project specific, so you can access the project's own feature
-    store but also any feature store which has been shared with the project you connect
-    to.
-
-    This class provides convenience classmethods accessible from the `hsfs`-module:
-
-    !!! example "Connection factory"
-        For convenience, `hsfs` provides a factory method, accessible from the top level
-        module, so you don't have to import the `Connection` class manually:
-
-        ```python
-        import hsfs
-        conn = hsfs.connection()
-        ```
-
-    !!! hint "Save API Key as File"
-        To get started quickly,
-        you can simply create a file with the previously created Hopsworks API Key and
-        place it on the environment from which you wish to connect to the Hopsworks
-        Feature Store.
-
-        You can then connect by simply passing the path to the key file when
-        instantiating a connection:
-
-        ```python hl_lines="6"
-            import hsfs
-            conn = hsfs.connection(
-                'my_instance',                      # Hostname of your Feature Store instance
-                443,                                # Port to reach your Hopsworks instance, defaults to 443
-                'my_project',                       # Name of your Hopsworks Feature Store project
-                api_key_file='featurestore.key',    # The file containing the API key generated above
-                hostname_verification=True)         # Disable for self-signed certificates
-            )
-            fs = conn.get_feature_store()           # Get the project's default feature store
-
-            # or
-
-            import hopsworks
-            project = hopsworks.login()
-            fs = project.get_feature_store()
-        ```
-
-    Clients in external clusters need to connect to the Hopsworks Feature Store using an
-    API key. The API key is generated inside the Hopsworks platform, and requires at
-    least the "project" and "featurestore" scopes to be able to access a feature store.
-    For more information, see the integration guides.
-
-    # Arguments
-        host: The hostname of the Hopsworks instance in the form of `[UUID].cloud.hopsworks.ai`,
-            defaults to `None`. Do **not** use the url including `https://` when connecting
-            programatically.
-        port: The port on which the Hopsworks instance can be reached,
-            defaults to `443`.
-        project: The name of the project to connect to. When running on Hopsworks, this
-            defaults to the project from where the client is run from.
-            Defaults to `None`.
-        engine: Which engine to use, `"spark"`, `"python"` or `"training"`. Defaults to `None`,
-            which initializes the engine to Spark if the environment provides Spark, for
-            example on Hopsworks and Databricks, or falls back on Hive in Python if Spark is not
-            available, e.g. on local Python environments or AWS SageMaker. This option
-            allows you to override this behaviour. `"training"` engine is useful when only
-            feature store metadata is needed, for example training dataset location and label
-            information when Hopsworks training experiment is conducted.
-        hostname_verification: Whether or not to verify Hopsworks' certificate, defaults
-            to `True`.
-        trust_store_path: Path on the file system containing the Hopsworks certificates,
-            defaults to `None`.
-        cert_folder: The directory to store retrieved HopsFS certificates, defaults to
-            `"/tmp"`. Only required when running without a Spark environment.
-        api_key_file: Path to a file containing the API Key, defaults to `None`.
-        api_key_value: API Key as string, if provided, `api_key_file` will be ignored,
-            however, this should be used with care, especially if the used notebook or
-            job script is accessible by multiple parties. Defaults to `None`.
-
-    # Returns
-        `Connection`. Feature Store connection handle to perform operations on a
-            Hopsworks project.
-    """
-
-    @not_connected
-    def connect(self):
-        """Instantiate the connection.
-
-        Creating a `Connection` object implicitly calls this method for you to
-        instantiate the connection. However, it is possible to close the connection
-        gracefully with the `close()` method, in order to clean up materialized
-        certificates. This might be desired when working on external environments such
-        as AWS SageMaker. Subsequently you can call `connect()` again to reopen the
-        connection.
-
-        !!! example
-            ```python
-            import hsfs
-            conn = hsfs.connection()
-            conn.close()
-            conn.connect()
-            ```
-        """
-        super().connect()
-
-        # init engine
-        engine.init(self._engine)
-
-        self._feature_store_api = feature_store_api.FeatureStoreApi()
-
-    @usage.method_logger
-    @connected
-    def get_feature_store(
-        self, name: Optional[str] = None
-    ) -> feature_store.FeatureStore:
-        """Get a reference to a feature store to perform operations on.
-
-        Defaulting to the project name of default feature store. To get a
-        Shared feature stores, the project name of the feature store is required.
-
-        !!! example "How to get feature store instance"
-
-            ```python
-            import hsfs
-            conn = hsfs.connection()
-            fs = conn.get_feature_store()
-
-            # or
-
-            import hopsworks
-            project = hopsworks.login()
-            fs = project.get_feature_store()
-            ```
-
-        # Arguments
-            name: The name of the feature store, defaults to `None`.
-
-        # Returns
-            `FeatureStore`. A feature store handle object to perform operations on.
-        """
-        if not name:
-            name = client.get_instance()._project_name
-        return self._feature_store_api.get(util.append_feature_store_suffix(name))
-
-    def close(self) -> None:
-        """Close a connection gracefully.
-
-        This will clean up any materialized certificates on the local file system of
-        external environments such as AWS SageMaker.
-
-        Usage is recommended but optional.
-
-        !!! example
-            ```python
-            import hopsworks
-            conn = hopsworks.connection()
-            conn.close()
-            ```
-        """
-        super().close()
-        engine.stop()
-        self._feature_store_api = None
